Dependency Injection
Injectable means that this service itself might have injected dependencies.
Component classes dont have @Injectable since they yhave the @Component decorator which informs angular that it might have dependecies and make use of the dependency injection system.

A dependency injector is a design pattern as well as a framework provided by Angular

After creating the service you need to register the service with a dependency injector.
There are multiple places where you can register the service. But where you register is important.
Mainly because Angular has a Hierarchical Dependency Injection.
To register a service, you need to providers metadata.
Next step is to declare the dependency in the components that require them.
This is done in the constructor of the receiving component.

https://www.udemy.com/angular-mastering-the-basics/learn/v4/t/lecture/10121078?start=0