Services Http and Observables
use the catch operator on the Observable. If there is an error, the error handler is executed. The components that have subscribed to the Observable have access to the data and the error.
If there is no error, the you get only the data. If there is an error , it comes in the error.

https://www.udemy.com/angular-mastering-the-basics/learn/v4/t/lecture/10121084?start=0