Routing
1. Create an angular project with routing option
2. Generate required components
3. Configure the routes.
4. Add buttons and use directives to navigate

Routing Steps :
In case the project was not created with the --routing option, it can be later created as
 1. Add  <base href="/"> in index.html
 2. Add the file app-routing-module.ts in app folder.
	This file contains the routing module for the app.
	This is where we configure the routes.
	Each route has a path and a component.
 3. Import app routing module in app.module.ts
	import { AppRoutingModule } from './app-routing.module';

Routes are created in app-routing.module.ts in the Routes array.
Since we import the routing components (DepartmentListComponent and EmployeeListComponent) here, we don't have to do it in app.module.ts
Then we create an array routingComponents in app-routing.module.ts
In the app.module.ts : we import this routingComponents array.

To navigate via buttons, use the routeLink directives
<nav>
    <button routerLink="/departments" routerLinkActive="active">Department</button>
    <button routerLink="/employees" routerLinkActive="active">Employees</button>
</nav>

https://www.udemy.com/angular-mastering-the-basics/learn/v4/t/lecture/10121086?start=0