Wildcard route : Commit message correction
Wildcard route is the route that is added at the end and is loaded if the user clicks on or navigates to a route that does not exist.

Redirection is when coming to one route should internally navigate to another route.

Best way to handle invalid routes is through the wildcard routes.
With this way, you have a page not found route.

The wild card route should always be the last route of the route config list.
The flow is that it matches the pattern from top to bottom.
So make sure that the routes are configured from the most specific to the least specific.

You an also redirect a route to another route.
This can be done using : redirectTo and pathMatch property of the route.

https://www.udemy.com/angular-mastering-the-basics/learn/v4/t/lecture/10121088?start=0