Parameterized Routes
You can create a route with a placeholer (parameter).
To pass a parameter while navigating, we make use of the router service.
It has a navigate method to which we can pass a link params array.

To read the param we make use of the activated route service.
We read a snapshot of the parameter and use the paramAPI.
This can then be used anyway we like.
eg: binding it to the view via interpolation

https://www.udemy.com/angular-mastering-the-basics/learn/v4/t/lecture/10121090?start=0