ParamMap Observable
drawback of snapshot approach :
Angular keeps track if it can reuse the existing component or does it need to initialize a new component.
If it is going to use the same component, the ngOnInit() method is not called.
Hence the snapshot does not get updated.

The paramMap Observable method is better.
this.route.paramMap.subscribe((params : ParamMap) => {
      let id = parseInt(params.get('id'));
      this.departmentId = id;
});

https://www.udemy.com/angular-mastering-the-basics/learn/v4/t/lecture/10121094?start=0